requirements.txt
flask
pdf2docx
python-docx
PyPDF2
gunicorn


